requirements.txt
apache-airflow==2.9.0
psycopg2-binary
pandas
matplotlib